requirements.txt
aiogram
python-dotenv
openpyxl
gspread
oauth2client